🚚 Move: components for TEAMS section to `$lib/components/home/teams` directory [#166]
