+++ src/lib/components/home/teams/GearsAndSettings.svelte
